feat: matched gradients
