+++ src/utils/utils.ts
@@ -35,3 +35,21 @@ export function formatLargeNumber(n: number | bigint) {
   }
   return num.toString();
 }
+
+export const getModelGradient = (modelFullName: string) => {
+  const gradients = [
+    "from-indigo-500 via-sky-500 to-violet-500",
+    "from-sky-500 via-emerald-500 to-teal-500",
+    "from-violet-500 via-rose-500 to-amber-500",
+    "from-rose-500 via-fuchsia-500 to-indigo-500",
+    "from-emerald-500 via-teal-500 to-cyan-500",
+    "from-amber-500 via-orange-500 to-red-500",
+    "from-fuchsia-500 via-purple-500 to-pink-500",
+    "from-blue-500 via-indigo-500 to-violet-500",
+  ];
+
+  const seed = modelFullName
+    .split("")
+    .reduce((acc, char, i) => acc + char.charCodeAt(0) * (i + 1), 0);
+  return gradients[seed % gradients.length];
+};
